feat: split workspace hooks up into different stores for efficiency
