+++ frontend/iou/src/hooks/workspace/useDimensions.ts
@@ -1,9 +1,23 @@
-import {useWorkspaceStore} from "@/hooks/workspace/useWorkspaceStore.ts";
 import {useShallow} from "zustand/react/shallow";
 import type {Dimensions} from "@/hooks/workspace/workspaceTypes.ts";
+import {create} from "zustand/react";
 
-export default function useDimensions(): [Dimensions, (dimensions: Dimensions | ((dimensions: Dimensions) => Dimensions)) => void] {
-  return useWorkspaceStore(
+type DimensionsMutator = (dimensions: Dimensions) => void;
+
+export interface DimensionsStore {
+  setDimensions: DimensionsMutator;
+  dimensions: Dimensions;
+}
+
+export const useDimensionsStore = create<DimensionsStore>((set) => ({
+  dimensions: "3d",
+  setDimensions: (dimensions: Dimensions) => {
+    set({ dimensions: dimensions });
+  },
+}))
+
+export default function useDimensions(): [Dimensions, DimensionsMutator] {
+  return useDimensionsStore(
     useShallow((state) => [state.dimensions, state.setDimensions]),
-  ) as [Dimensions, (dimensions: Dimensions | ((dimensions: Dimensions) => Dimensions)) => void]
+  ) as [Dimensions, DimensionsMutator];
 }
