+++ frontend/iou/src/hooks/workspace/workspaceTypes.ts
@@ -7,25 +7,16 @@ export type Vec2 = [number, number, 0];
 // The return type for useWorkspace
 export interface WorkspaceMetadata {
   name: string;
-  dimensions: Dimensions;
 }
 
 // The data type held for each shape in the workspace store
-export interface Shape {
+export interface ShapeData {
   vertices: number[][];
   faces: number[][];
 
   // Extra data for managing workers calculating faces using the convex hull algorithm
   isPending: boolean;
   currentJobId?: number;
-}
-
-// The return type for useWorkspaceStore
-export interface WorkspaceState extends WorkspaceMetadata {
-  setDimensions: (dimensions: Dimensions) => void;
-
-  shapes: Map<string, Shape>;
-  setVertices: (id: string, vertices: number[][]) => void;
 }
